+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Improved sound options: you can choose the channel the sound plays on (Master, Music, SFX, etc), and you can ask it to play while the game isn't focused / the channel is muted
 * Stopped hiding chat-channel output from the config for other addons using LibSink
+* Allow right-clicking on the target popup to close it
 * Fixed various quest ids
 * Background Shadowlands work